Use a dedicated `coin` field within `TokenBundle` to store the ada quantity.
In this change, we:

1. Use a dedicated `coin` field within `TokenBundle` to store the ada `Coin`
   amount.

2. Extract out the map part of `TokenBundle` into a separate submodule.
   The map has an invariant that must be strictly adhered to, so it makes
   sense to limit the size of its interface (so as to limit the number of
   ways in which the invariant could be violated) and test it in isolation.

The interface of `TokenBundle` is more-or-less preserved as-is.
