+++ lib/core/cardano-wallet-core.cabal
@@ -165,6 +165,7 @@ library
       Cardano.Wallet.Primitive.Types.Hash
       Cardano.Wallet.Primitive.Types.RewardAccount
       Cardano.Wallet.Primitive.Types.TokenBundle
+      Cardano.Wallet.Primitive.Types.TokenBundle.TokenMap
       Cardano.Wallet.Primitive.Types.TokenPolicy
       Cardano.Wallet.Primitive.Types.TokenQuantity
       Cardano.Wallet.Primitive.Types.Tx
@@ -324,6 +325,7 @@ test-suite unit
       Cardano.Wallet.Primitive.TypesSpec
       Cardano.Wallet.Primitive.Types.HashSpec
       Cardano.Wallet.Primitive.Types.TokenBundleSpec
+      Cardano.Wallet.Primitive.Types.TokenBundle.TokenMapSpec
       Cardano.Wallet.Primitive.Types.TokenQuantitySpec
       Cardano.Wallet.RegistrySpec
       Cardano.Wallet.TransactionSpec
