+++ lib/core/test/data/Cardano/Wallet/Primitive/Types/TokenQuantity.json
@@ -1,15 +1,15 @@
 {
-    "seed": -2492658773060386009,
+    "seed": 6142827867234780355,
     "samples": [
-        25,
-        -1,
-        0,
         -1.0e1000,
-        3.034856222810095003090312922436194634318063577209271644140789794598391154732016710318304669572999972380688110441960965670157578300920031818299245693164119312501752727831349481636798503731847103706789985445615699384395516119574136230145588568564573610177246003033792507229377890587706901767887179067978131856413898978965218480613554538720525806683961365906440504706488176022857171948702930934139443165117811016484064377109069488244052834966597457120095163612120837800766753448235015697107770618974507283724741713251784680215728904066982252672930711185186037184927223959789333800825640596680558836395597583085561591108033979417511082531916254267526087726198412924976559015513782752681330670010930658052601294417156169526306202011117389253375145811135484227203942114449662961489280438981309196933434999375968226632135920289432972747414886478075066220164875439873339560055003279170797845190520404446981447987396750951123175753891720972345838035516828165402368385958483138775541730961866157102483846921282e999,
-        -15,
-        -86,
+        13,
         -1,
-        -1,
-        1
+        -63,
+        -98,
+        4.480180432846184940432227313634794296403314975864796624878797151928710234030883532096928606359711270793533750532982885991505413717458374397983207502642250465112055712853439709385793062236306649252986035625857475909251553750040511689186123609295886686449451817326537072201413436137354527707978149195588253425702601021971933233820738687036095053536642019482038690873818622087976475603709178381264253850624018829076907976858589603585184863043679315492475888869516428706864000264103064120367778341916984451442921799070429554142410462015648787874696808774035262196504086753686919493193015930726123500529509962189169886730112874910747499380909668343849035934786921519228964266419754224093540002140378914525257454236091394777253626694700950267537706406425052785756782298091918229948577431203671466392624131083468992969224335562779617450384981197355658676727137633814545483571215386871218016184728828805994548950010359155880951411838469583890398443991745848220086987691138173255655597165233779203239282329994e999,
+        1.0e1000,
+        0,
+        5.046977952649305764758720817857783252374744499577087711555230104454856534253386830929191447228622880219911505629947331593725557622399317952218535090147249145964836280285454503036645413140853100042315346780606193469342912321222573924631994157933759790758685578378560504438691309373424645085807864941401993270405658886174229534620825617441998585779721917037892905199918571835003554031020670886892041443280270395734528073137614788767745469705770099950990721833926808166706944541881778443394387384881716043054618405589101119070745099709213690994812799382749355054093854842775376655757079671575056029344683895339980441388566736614112945547682347910073489194598048216631242779010010811817836866653671930232350379533882784835589550791231496472887548846651617864318299632860975234521403732176229642211928731586678615047762305582939062680080482587161394987793815739200205713901490596676481681078986961871900431122560834866861840441159217070995367033598314944513773449264258293674402754532043211713848467052221e999,
+        8.624216573584167714977299418639142242201248601587951627380182168964848573061716676350841135362193296635316181403456706357840170038305576116226557633617741557217676492227251555847444260487202432856651356482170659083732288110414181338484993816031745756087415931553796922614484790028959793722317003885055240615594093272033069504180559000274487192261685201277566704093118381722634091535942524506206959408057255076616077149718446360103215587549343258434300632836456487006120885097617032238294687520875677659713376577916225145463953990041466716819348157393667396421274315622290851392302481840879968490232167975842686376955351938158982525796447542061210207344125886075912982318005458302215322630885734521308223479876473793149047453286314912049958734989441299864644766709380027427982017449631982248842819198100010124718759567492844267774050682769369387901269250274941595473611497999073277535902065169915597261490971189133185309494075549266402135277590178243786754233441198373534230822994590617409267637804004e999
     ]
 }
